% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{coerce_numeric_to_sparse_numeric}
\alias{coerce_numeric_to_sparse_numeric}
\title{Coerce numeric to sparse_numeric}
\arguments{
\item{from}{A numeric vector.}
}
\value{
An object of class \code{sparse_numeric}.
}
\description{
Converts a dense numeric vector to a \code{sparse_numeric} object by
storing only the non-zero entries.
}
