% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{coerce_sparse_to_numeric}
\alias{coerce_sparse_to_numeric}
\title{Coerce sparse_numeric to numeric}
\arguments{
\item{from}{A \code{sparse_numeric} object.}
}
\value{
A numeric vector of length \code{from@length}.
}
\description{
Expands a \code{sparse_numeric} object to its full dense numeric
representation.
}
