% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{nnzero}
\alias{nnzero}
\alias{nnzero,sparse_numeric-method}
\title{Number of non-zero entries}
\usage{
nnzero(x)

\S4method{nnzero}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}
}
\value{
An integer count of non-zero entries.
}
\description{
Returns the number of non-zero entries in a \code{sparse_numeric}
vector.
}
