% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{norm,sparse_numeric,ANY-method}
\alias{norm,sparse_numeric,ANY-method}
\title{Norm of a sparse_numeric vector}
\usage{
\S4method{norm}{sparse_numeric,ANY}(x, type, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
A numeric scalar giving the Euclidean norm of \code{x}.
}
\description{
Computes the Euclidean norm of a \code{sparse_numeric} vector, defined
as the square root of the sum of squared entries.
}
