% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm of a vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric scalar giving the norm of \code{x}.
}
\description{
Generic function for computing the norm of an object.
The method for \code{sparse_numeric} computes the Euclidean norm,
defined as the square root of the sum of squared entries.

Computes the Euclidean norm of a \code{sparse_numeric} vector, defined
as the square root of the sum of squared entries.
}
