% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot overlapping non-zero entries of two sparse_numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}

\item{...}{Additional arguments passed on to \code{plot()}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plots the values of two \code{sparse_numeric} vectors at positions
where they both have non-zero entries.
}
