% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{+,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse_numeric vectors elementwise with \code{*}}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\value{
A \code{sparse_numeric} object representing \code{x * y}.
}
\description{
This is an operator alias for \code{sparse_mult}, allowing you to write
\code{x * y} when both are \code{sparse_numeric} objects.
}
