% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Show method for sparse_numeric}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\description{
Displays a compact representation of a \code{sparse_numeric} object.
}
