% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse_numeric vectors elementwise}
\usage{
sparse_mult(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}
}
\value{
A \code{sparse_numeric} object representing \code{x * y}.
}
\description{
Computes the elementwise product of two \code{sparse_numeric} vectors
that have the same length.
}
