% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse_numeric vectors}
\usage{
sparse_sub(x, y)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}
}
\value{
A \code{sparse_numeric} object representing \code{x - y}.
}
\description{
Computes the elementwise difference of two \code{sparse_numeric}
vectors that have the same length.
}
