% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_DataProduct.R
\name{+,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Add two sparse_numeric vectors with \code{+}}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{First \code{sparse_numeric} vector.}

\item{e2}{Second \code{sparse_numeric} vector.}
}
\value{
A \code{sparse_numeric} object representing \code{e1 + e2}.
}
\description{
This is an operator alias for \code{sparse_add}, allowing you to write
\code{e1 + e2} when both are \code{sparse_numeric} objects.
}
